fix all the warnings

+++ rubikomo/formula.cpp
@@ -10,7 +10,7 @@ Formula::Formula(std::string formula)
 
    int stepindex = 0;
    int charindex = 0;
-   int length = formula.length();
+   size_t length = formula.length();
    while (stepindex < MAX_STEPS && charindex < length)
    {
        cStep[0] = formula[charindex++];
